+++ Libraries/SPTushonka.Server.Core/Models/Eft/Game/GetItemPricesResponse.cs
@@ -6,7 +6,7 @@ namespace SPTarkov.Server.Core.Models.Eft.Game;
 public record GetItemPricesResponse
 {
     [JsonPropertyName("supplyNextTime")]
-    public double? SupplyNextTime { get; set; }
+    public long? SupplyNextTime { get; set; }
 
     [JsonPropertyName("prices")]
     public Dictionary<MongoId, double>? Prices { get; set; }
